Move project/worksapce parsing to a seperate file
